Added some transwiki links
Added a few commonly used transwiki links (zhdel is not used in wikimedia sites, but it would be quite useful when discussing the content of a deleted page). Phabricator and OTRS do not follow the regular rule (domain/wiki/$1) thus a new array and a variable were used. Thank you for your time in reviewing this patch.